Corrections for symmetry factor feature in GRChombo
Corrected the params files and any uses of AMRReductions in
both of the Examples, to account for the integral symmetry factor
feature in GRChombo (see GRChombo Issue No. 197). Specifically:

1. Added the symmetry_correction flag to the params files
(under the Boundary Condition parameters)
2. Provided the chombo parameter storing the symmetry factor
to every instance of the norm and sum AMRReductions functions,
in the specificPostTimeStep() function for each example.

Note that the symmetry factor is calculated entirely in the
ChomboParameters.h file of the main GRChombo repo.