Corrections for symmetry factor feature in GRChombo (#26)
* Corrections for symmetry factor feature in GRChombo

Corrected the params files and any uses of AMRReductions in
both of the Examples, to account for the integral symmetry factor
feature in GRChombo (see GRChombo Issue No. 197). Specifically:

1. Added the symmetry_correction flag to the params files
(under the Boundary Condition parameters)
2. Provided the chombo parameter storing the symmetry factor
to every instance of the norm and sum AMRReductions functions,
in the specificPostTimeStep() function for each example.

Note that the symmetry factor is calculated entirely in the
ChomboParameters.h file of the main GRChombo repo.

* Corrections for symmetry factor feature in GRChombo

Added the symmetry factor flag to params for both examples,
enabling calls to AMRReductions the ability to correct for
symmetry factor associated with reflective boundary conditions
(see Issue 197).

* Removed symmetry factor from amr_reductions

* Removed symmetry_correction flag

* Update intel test, as in GRChombo

* Small changes in tests

* Revert "Small changes in tests"

This reverts commit ac53bf2b941322664d2d83a3c2a021352e304d06.

* Small changes to tests

* Remove output path in aparams and symmetry factors in conservation plots

* Remove mpi matrix

* typo fix

* Comment out max_steps line

---------

Co-authored-by: dinatraykova <dina.traykova@aei.mpg.de>